Temporarily use Float64 for some tests due to Enzyme issue

+++ test/enzyme/eigh.jl
@@ -3,7 +3,9 @@ using Test
 using LinearAlgebra: Diagonal
 using CUDA, AMDGPU
 
-BLASFloats = (Float32, ComplexF64) # full suite is too expensive on CI
+# infinity-norm doesn't play nicely with Float32, Enzyme, and 1.12
+# see https://github.com/EnzymeAD/Enzyme.jl/issues/2985
+BLASFloats = (Float64, ComplexF64) # full suite is too expensive on CI
 GenericFloats = ()
 @isdefined(TestSuite) || include("../testsuite/TestSuite.jl")
 using .TestSuite
